put cargo invoice generation into its own file
